Move error struct to its own file.
