+++ src/lib.rs
@@ -16,3 +16,4 @@ extern crate dirs;
 
 pub mod format;
 pub mod pki;
+pub mod error;
